+++ experiments/ClimaEarth/run_moist_held_suarez.jl
@@ -39,9 +39,6 @@ pkg_dir = pkgdir(ClimaCoupler)
 ## helpers for component models
 include("components/atmosphere/climaatmos.jl")
 
-## helpers for user-specified IO
-include("user_io/user_logging.jl")
-
 #=
 ### Setup simulation parameters
 Here we follow Thatcher and Jablonowski (2016).
@@ -214,7 +211,7 @@ dates = (; date = [date], date0 = [date0], date1 = [Dates.firstdayofmonth(date0)
 
 checkpoint_cb = TimeManager.HourlyCallback(
     dt = FT(480),
-    func = checkpoint_sims,
+    func = Checkpointer.checkpoint_sims,
     ref_date = [dates.date[1]],
     active = hourly_checkpoint,
 )